refactor(web): split viewer.js into modules; replace Tailwind CDN with a real build
Task 1 — split web/js/viewer.js (944 lines) into ES modules:
- web/js/resourceChart.js: renderResources + resourcesUrl/stageLabel/uPlot config.
- web/js/windowControl.js: makeWindowControl + winControls and its private helpers.
- viewer.js now imports and coordinates them.
submission.html already loads viewer.js as <script type="module">, so this uses
ES-module imports between siblings (no HTML change needed). app.js's window.QSM
globals and the uPlot IIFE global stay reachable from module scope, as before.

Task 2 — replace https://cdn.tailwindcss.com with a compiled stylesheet:
- web/package.json (tailwindcss@3.4 + build:css), web/tailwind.config.js
  (darkMode: class, Inter font, content globs over *.html + js/*.js, safelist
  guard for state/badge palettes), web/src/input.css (@tailwind directives).
- Built to web/styles.css (minified, committed) via the Tailwind CLI.
- Every HTML page now links styles.css instead of the CDN dev build; the inline
  tailwind.config scripts are removed (config lives in tailwind.config.js).
- pages.yml gains a Node setup + `npm ci` + `npm run build:css` step so CI
  regenerates styles.css before deploying.
