+++ .github/workflows/pages.yml
@@ -36,6 +36,19 @@ jobs:
     steps:
       - uses: actions/checkout@v7
 
+      - name: Set up Node
+        uses: actions/setup-node@v4
+        with:
+          node-version: "20"
+
+      # Compile Tailwind to web/styles.css before assembling the site (replaces the old CDN JIT build).
+      # The committed styles.css is regenerated here so a push always deploys CSS matching current markup.
+      - name: Build Tailwind CSS
+        working-directory: web
+        run: |
+          npm ci
+          npm run build:css
+
       - name: Assemble site (web/ + results/ served alongside)
         run: |
           mkdir -p _site
