+++ web/src/input.css
@@ -1,0 +1,5 @@
+/* Tailwind entry point. Built to ../styles.css by `npm run build:css` (see package.json).
+   Replaces the old CDN JIT build (https://cdn.tailwindcss.com). */
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
