+++ web/submission.html
@@ -6,8 +6,7 @@
   <title>QSM-CI — Submission</title>
   <link rel="preconnect" href="https://fonts.googleapis.com" />
   <link href="https://fonts.googleapis.com/css2?family=Inter:wght@400;500;600;700;800&display=swap" rel="stylesheet" />
-  <script src="https://cdn.tailwindcss.com"></script>
-  <script>tailwind.config = { darkMode: "class", theme: { extend: { fontFamily: { sans: ["Inter", "system-ui", "sans-serif"] } } } };</script>
+  <link rel="stylesheet" href="styles.css" />
   <script src="js/app.js"></script>
   <style>
     /* dual-range windowing slider with value bubbles + intensity histogram */
